pyproject.toml
[tool.poetry]
name = "bibbi-aut"
version = "0.1.0"
description = "Conversion scripts for Biblioteksentralen's authority records"
authors = ["Dan Michael O. Heggø <Dan.Michael.Heggo@bibsent.no>"]
license = "MIT"
packages = [
    { include = "bibbi" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.0.3"
openpyxl = "^3.0.3"
rdflib = "^5.0.0"
pyyaml = "^5.3.1"
rtyaml = "^1.0.0"
psutil = "^5.7.0"
otsrdflib = "^0.5.0"
pyodbc = "^4.0.30"
pydash = "^4.7.6"
feather-format = "^0.4.1"
skosify = {git = "https://github.com/NatLibFi/Skosify.git", rev = "master"}
python-dotenv = "^0.13.0"
bonobo = "^0.6.4"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"
pytest-testdox = "^1.2.1"
ipykernel = "^5.3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
